+++ example/gemini-client/bin/main.dart
@@ -17,7 +17,7 @@ void main(List<String> args) async {
   }
 
   final client = GoogleMcpClient(
-    GenerativeModel(model: 'gemini-2.5-pro-exp-03-25', apiKey: apiKey),
+    GenerativeModel(model: 'gemini-2.0-flash', apiKey: apiKey),
     Client(Implementation(name: "gemini-client", version: "1.0.0")),
   );